Updated photometric calibration scirpt
The new 08_photometricCalibration script is divided into part a)
selection of good photometric stars, and b) execution of photometric
calibration.

The user is encouraged to be semi-liberal in their selection of "good"
photometric stars. If one star in the USNO-B1.0 catalog appears to have
two separate entries in a given USNO-B1.0 epoch, the code will determine
which of the two entries  is better suited for a linear fit based on the
probability that the entry is drawn from the outlier distribution of the
MCMC linear fit.