Added "background-count-model" subtraction
Step 05_avgPolAngImages.py script now estimates the background level
using the "off-target" images and fits a smoothly varying background
level vs. time. This background level model is then interpolated to the
"on-target" times, and that value is used to scale up an average,
normalized, background image, which is then subtracted from the
corresponding "on-target" image.