Use COMPLETE list of Whittet stellar data
The Whittet et al. (1992) list of polarization measurements and stellar
properties are reported in full within this CSV.

There may be some errors, so this needs to be double checked for other
stars, but at least this will work such that any given star from this
list can be used as a polarization standard calibrator.